requirements.txt
django
pyusb==1.0.0a3
